refactor(graph): move resolvers into subpackage
