+++ gitstore-api/internal/graph/resolver/query_helpers.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 // Copyright (c) 2026 GitStore contributors
 
-package graph
+package resolver
 
 import (
 	"context"
