+++ gitstore-api/internal/graph/resolver/schema.resolvers.go
@@ -1,4 +1,4 @@
-package graph
+package resolver
 
 // This file will be automatically regenerated based on the schema, any resolver
 // implementations
